requirements.txt
Flask
flask-cors
openai
python-dotenv
streamlit
requests